Separating views.py to files, kubernetes configuration
Moved functions from views.py to code_change_views.py and search_views.py. Changed __init__.py for views. Added elasticsearch and kibana configuration files. Added plantuml model.
