+++ kubernetes_config/kibana-configmap.yaml
@@ -1,0 +1,21 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: kibana-config-map
+data:
+    kibana.yml: |
+      server.port: 5601
+      server.host: "0.0.0.0"
+      server.basePath: ""
+      elasticsearch.url: "http://elasticsearch:9200"
+      elasticsearch.username: "elastic"
+      elasticsearch.password: "changeme"
+      kibana.index: ".kibana"
+      kibana.defaultAppId: "discover"
+      elasticsearch.ssl.verificationMode: none
+      elasticsearch.pingTimeout: 1500
+      elasticsearch.requestTimeout: 30000
+      logging.dest: stdout
+      logging.verbose: false
+      ops.interval: 5000
+      i18n.defaultLocale: "en"
