+++ uks_app/views/__init__.py
@@ -3,3 +3,5 @@ from .issue_views import *
 from .subissue_views import *
 from .label_views import *
 from .other_views import *
+from .search_views import *
+from .code_change_views import *
